+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Migrating the existing HTML website to a modern React-based architecture for improved performance, scalability, and maintainability.
 - View a complete list of all registered websites in the database.
 - A dedicated page to review and verify all data registered by SteamShark.
+- Added README.md
+
+--- 
+
+- [Commit rework website](https://github.com/steamShark/website/commit/a0cdf7a5746e899e1d9dfd6116eea76327646834)
 
 ## [1.0.0] - 19/01/2025